+++ Common/CommandAvailablity.cs
@@ -4,7 +4,7 @@
 // This source code is licensed under the BSD-style license found in the
 // LICENSE file in the root directory of this source tree.
 
-namespace eVolve.CsvDataExchange.Revit;
+namespace eVolve.ExtensionsCommon.Revit;
 
 /// <summary> Determines if a Revit document is currently loaded. </summary>
 internal class CommandAvailability : IExternalCommandAvailability
