+++ eVolveMEP CSV Data Exchange/Application.cs
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) 2023 eVolve MEP, LLC
 // All rights reserved.
-//
+// 
 // This source code is licensed under the BSD-style license found in the
 // LICENSE file in the root directory of this source tree.
 
@@ -31,7 +31,7 @@ public class ApplicationMechanical : IExternalApplication
         var ribbonButton = eVolve::eVolve.Core.Revit.Integration.API.CreateButton(Resources.ButtonText,
             System.Reflection.Assembly.GetExecutingAssembly().Location,
             typeof(Command),
-            typeof(CommandAvailability),
+            typeof(ExtensionsCommon.Revit.CommandAvailability),
             BitmapFrame.Create(Command.IconResource),
             string.Format(Resources.ToolTipText, HostProductName),
             Command.HelpLinkUrl);
