+++ eVolveMEP Revit Extensions.sln
@@ -14,6 +14,12 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "eVolveMEP Data Table Tools", "eVolveMEP Data Table Tools\eVolveMEP Data Table Tools.csproj", "{85DE10A1-786E-45EF-992A-80E3815A5A4D}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Common", "Common", "{B15BF223-593D-409C-B1EB-DBEDE497501A}"
+	ProjectSection(SolutionItems) = preProject
+		Common\CommandAvailablity.cs = Common\CommandAvailablity.cs
+		Common\Methods.cs = Common\Methods.cs
+	EndProjectSection
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
